event_management/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers
